Fix the Screenshot test filter
Class.forName("com.gravatar.app.testUtils.roborazzi.ScreenshotTest")

wasn't working as expected. The previous approach was trying to avoid any module using the GravatarComposeConventionPlugin to also depend on "testUtils".

However, it makes sense that if it's a compose module, it likely needs screenshot tests, so it sounds reasonable to require that dependency.